+++ module_20/composer.json
@@ -1,0 +1,20 @@
+{
+    "name": "module/twenty",
+    "type": "project",
+    "license": "MIT",
+    "autoload": {
+        "psr-4": {
+            "Module\\Twenty\\": "src/classes"
+        }
+    },
+    "authors": [
+        {
+            "name": "vladSHOKO",
+            "email": "kuzia_vladik@mail.ru"
+        }
+    ],
+    "require": {},
+    "require-dev": {
+        "phpunit/phpunit": "^11.0"
+    }
+}
