+++ module_20/data/authorization.php
@@ -1,5 +1,7 @@
 <?php
 
+namespace Module\Twenty;
+
 if (isset($_POST['login']) && isset($_POST['password'])) {
     $user = new Authorization(
         '127.0.0.1',
