+++ module_20/route/userProfile.php
@@ -1,5 +1,7 @@
 <?php
 
+namespace Module\Twenty;
+
 $user = new User('127.0.0.1', 'root', 'Faraonkill1', 'authorization');
 foreach ($user->getData() as $key => $item) {
     ?> <p class="author__name"><?php
