+++ module_20/src/core.php
@@ -214,160 +214,3 @@ function updateLoginCookie(): void
     }
 }
 
-class Database
-{
-    public $hostname;
-    public $username;
-    public $password;
-    public $dbname;
-
-    public function __construct($hostname, $username, $password, $dbname)
-    {
-        $this->hostname = $hostname;
-        $this->username = $username;
-        $this->password = $password;
-        $this->dbname = $dbname;
-    }
-
-    public function databaseConnect(): object
-    {
-        return new PDO(
-            "mysql:host=$this->hostname;dbname=$this->dbname;charset=utf8",
-            $this->username,
-            $this->password
-        );
-    }
-}
-
-class Authorization extends Database
-{
-    public function validateUser(): void
-    {
-        $user = $this->findUsersForLogin();
-        $this->setDataInSession($user, $this->checkUserPassword($user));
-    }
-
-    private function findUsersForLogin(): array
-    {
-        $connect = parent::databaseConnect();
-        $request = $connect->prepare(
-            'select id, login, password from users where login = :login'
-        );
-        $request->execute(['login' => $_POST['login']]);
-        $result = $request->fetch(PDO::FETCH_ASSOC);
-        var_dump($request->fetch(PDO::FETCH_ASSOC));
-        if (!$result) {
-            return [];
-        } else {
-            return $result;
-        }
-    }
-
-    private function checkUserPassword(array $user): bool
-    {
-        return password_verify(
-            $_POST['password'],
-            $user['password']
-        );
-    }
-
-    private function setDataInSession(array $user, bool $correctPassword): void
-    {
-        if (!empty($user) && $correctPassword) {
-            $_SESSION['auth'] = true;
-            $_SESSION['user_id'] = $user['id'];
-        }
-    }
-}
-
-class User extends Database
-{
-    public function getData(): array
-    {
-        $connect = parent::databaseConnect();
-        $request = $connect->prepare(
-            'select name as Имя, surname as Фамилия, father_name as Отчество, email as Почта, phone_number as Телефон from users where id = :id'
-        );
-        $request->execute(['id' => $_SESSION['user_id']]);
-        return $request->fetch(PDO::FETCH_ASSOC);
-    }
-}
-
-class Mail extends Database
-{
-    public function getMessageList(): array
-    {
-        $connect = parent::databaseConnect();
-        $request = $connect->prepare(
-            'select * from messages where recipient_id = :id'
-        );
-        $request->execute(['id' => $_SESSION['user_id']]);
-        $messageList = [];
-        while ($row = $request->fetch(PDO::FETCH_ASSOC)) {
-            $messageList[$row['id']] = $row;
-        }
-        return $messageList;
-    }
-
-    public function getDataOfCurrentMessage(): array
-    {
-        $connect = parent::databaseConnect();
-        $request = $connect->prepare(
-            'select * from messages where id = ? and recipient_id = ?'
-        );
-        $request->execute([$_GET['id'], $_SESSION['user_id']]);
-        return $request->fetch(PDO::FETCH_ASSOC);
-    }
-
-    public function getDataOfSender(array $message): array
-    {
-        $connect = parent::databaseConnect();
-        $request = $connect->prepare(
-            'select name, surname, email from users where id = ?'
-        );
-        $request->execute([$message['sender_id']]);
-        return $request->fetch(PDO::FETCH_ASSOC);
-    }
-
-    public function updateMessageStatus(): void
-    {
-        $connect = parent::databaseConnect();
-        $requestOfUpdateStatus = $connect->prepare(
-            'update messages set readed = 1 where id = ?'
-        );
-        $requestOfUpdateStatus->execute([$_GET['id']]);
-    }
-
-    public function getListOfRecipients(): array
-    {
-        $connect = parent::databaseConnect();
-        $request = $connect->prepare(
-            'select id, name, surname from users order by name'
-        );
-        $request->execute();
-        $list = [];
-        while ($row = $request->fetch(PDO::FETCH_ASSOC)) {
-            $list[] = $row;
-        }
-        return $list;
-    }
-
-    public function insertMessage(): void
-    {
-        $connect = parent::databaseConnect();
-        if (isset($_POST['recipient'])) {
-            $request = $connect->prepare(
-                'insert into messages (title, text, sender_id, recipient_id, time) values (?, ?, ?, ?, ?)'
-            );
-            $request->execute(
-                array(
-                    $_POST['message_title'],
-                    $_POST['message_text'],
-                    $_SESSION['user_id'],
-                    $_POST['recipient'],
-                    date('Y-m-d H:i:s')
-                )
-            );
-        }
-    }
-}
